Move resolution to base class
